.github/scripts/requirements.txt
requests==2.24.0
python-gitlab==2.5.0
PyGithub==1.55
